requirements.txt
asgiref==3.7.2
certifi==2023.11.17
distlib==0.3.8
Django==5.0.1
django-api==0.1.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
pillow==10.2.0
pipenv==2023.11.17
platformdirs==4.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pygame==2.5.2
PyJWT==2.8.0
pytz==2024.1
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4
virtualenv==20.25.0
